% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{testNormalPlot}
\alias{testNormalPlot}
\title{Normal distribution plot}
\usage{
testNormalPlot(av1, sd1)
}
\arguments{
\item{av1}{mean value}

\item{sd1}{standard deviation}
}
\value{
Average and Standard Deviation
}
\description{
\code{testNormalPlot} Plot normal distribution
}
\examples{
# testNormalPlot(50, 10)
}
